Switch to push-based FlatGeobufWriter, add GeozeroRecordBatch
