+++ Cargo.toml
@@ -74,6 +74,7 @@ rstar = "0.12.2"
 serde = "1"
 serde_json = "1"
 serde_with = "3"
+tempfile = "3"
 thiserror = "1"
 tokio = { version = "1.9", default-features = false }
 url = "2.5"
