+++ rust/geoarrow-array/src/geozero/export/mod.rs
@@ -4,4 +4,5 @@ mod array;
 mod data_source;
 pub(crate) mod scalar;
 
+pub use data_source::GeozeroRecordBatch;
 pub use data_source::GeozeroRecordBatchReader;
